refactor(metrics): extract MetricKindCollection into kind_collection.rs
